+++ releaseNotes.md
@@ -12,6 +12,7 @@ edge / 2019-12-08
 - Breaking: removed executeSync method on Core. Up to the grammar to define it's own execute methods.
 - Breaking: removed default execute implementation on Core. Up to grammar to specify.
 - Breaking: jtree.makeProgram is now jtree.compileGrammarAndCreateProgram
+- Breaking: jtree.formatFile is now jtree.formatFileInPlace and jtree.formatProgram is now jtree.formatCode
 
 48.1.0 / 2019-12-06
 ===================